shipment_service/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.1.13
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.5.0
idna==3.7
mysql-connector-python==8.3.0
pillow==10.2.0
psycopg2-binary==2.9.9
pymongo==3.12.3
python-dotenv==1.0.1
pytz==2023.4
requests==2.32.1
sqlparse==0.2.4
tzdata==2024.1
urllib3==2.2.1